refactor(ingestion): Reorg to align to convention
